Better styling to line chart
